+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Master
+
+#### Breaking
+
+#### Enhancements
+
+#### Bug Fixes
+
 ## 3.0.0
 
 ##### Breaking